# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankalpa Odisha API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /announcements/{id}/attachments:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [Announcements]
      operationId: listAnnouncementAttachments
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AttachmentListResponse" } } } }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
    post:
      tags: [Announcements]
      operationId: registerAnnouncementAttachment
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RegisterAttachmentRequest" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/AttachmentResponse" } } } }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
  /attachments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    delete:
      tags: [Announcements]
      operationId: deleteAttachment
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
...
  /uoi-notes/{id}/attachments:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [UoiNotes]
      operationId: listUoiNoteAttachments
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/UoiNoteAttachmentListResponse" } } } }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
    post:
      tags: [UoiNotes]
      operationId: registerUoiNoteAttachment
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RegisterAttachmentRequest" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/UoiNoteAttachmentResponse" } } } }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
...
components:
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    AttachmentRecord:
      type: object
      properties:
        id: { type: integer }
        announcementId: { type: integer }
        objectPath: { type: string }
        fileName: { type: string }
        fileSize: { type: integer }
        contentType: { type: string }
        uploadedBy: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, announcementId, objectPath, fileName, fileSize, contentType, uploadedBy, createdAt]
    AttachmentListResponse:
      type: object
      properties:
        attachments:
          type: array
          items: { $ref: "#/components/schemas/AttachmentRecord" }
      required: [attachments]
    AttachmentResponse:
      type: object
      properties:
        attachment: { $ref: "#/components/schemas/AttachmentRecord" }
      required: [attachment]
    RegisterAttachmentRequest:
      type: object
      properties:
        objectPath: { type: string }
        fileName: { type: string }
        fileSize: { type: integer }
        contentType: { type: string }
        uploadToken:
          type: string
          description: >
            Server-issued one-time token from POST /storage/uploads/request-url
            that binds the upload to this actor + objectPath. Required.
      required: [objectPath, fileName, fileSize, contentType, uploadToken]
...
    UoiNoteAttachmentRecord:
      type: object
      properties:
        id: { type: integer }
        noteId: { type: integer }
        objectPath: { type: string }
        fileName: { type: string }
        fileSize: { type: integer }
        contentType: { type: string }
        uploadedBy: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, noteId, objectPath, fileName, fileSize, contentType, uploadedBy, createdAt]
    UoiNoteAttachmentListResponse:
      type: object
      properties:
        attachments:
          type: array
          items: { $ref: "#/components/schemas/UoiNoteAttachmentRecord" }
      required: [attachments]
    UoiNoteAttachmentResponse:
      type: object
      properties:
        attachment: { $ref: "#/components/schemas/UoiNoteAttachmentRecord" }
      required: [attachment]
...